Move error code utils into new file
Error code utils can be used for aggregating error
responses from multiples responses including use case
for managing separate route handlers for OEM namespace

Tests
Units tests are passing.
