+++ redfish-core/include/error_message_utils.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include <nlohmann/json.hpp>
 
 #include <string_view>
